refactor: address mr-tz review on modular layout/tags split
Move string extraction out of layout/types into layout/extract, fold
merge_overlapping_ranges into ranges, and consolidate layout-derived tag
checks into tags/checks. Remove the unused enrich shim, add pipeline.py as
the iteration-2 enrichment placeholder, restore script usage docs, and
clarify interim document vs results schemas with docstrings.
